requirements.txt
gradio>=3.50
pypdf
sentence-transformers
faiss-cpu
transformers
torch
numpy
